Change ID attribute from int to long

+++ src/main/java/io/datajek/moneytransferrest/repository/TransactionRepository.java
@@ -3,5 +3,5 @@ package io.datajek.moneytransferrest.repository;
 import io.datajek.moneytransferrest.model.TransactionEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TransactionRepository extends JpaRepository<TransactionEntity, Integer> {
+public interface TransactionRepository extends JpaRepository<TransactionEntity, Long> {
 }
